Remove commit() method Firehose
