+++ indexing-service/src/test/java/org/apache/druid/indexing/overlord/TaskLifecycleTest.java
@@ -291,19 +291,6 @@ public class TaskLifecycleTest
         }
 
         @Override
-        public Runnable commit()
-        {
-          return new Runnable()
-          {
-            @Override
-            public void run()
-            {
-
-            }
-          };
-        }
-
-        @Override
         public void close()
         {
 
@@ -343,19 +330,6 @@ public class TaskLifecycleTest
         public InputRow nextRow()
         {
           return inputRowIterator.next();
-        }
-
-        @Override
-        public Runnable commit()
-        {
-          return new Runnable()
-          {
-            @Override
-            public void run()
-            {
-
-            }
-          };
         }
 
         @Override
