+++ server/src/main/java/org/apache/druid/segment/realtime/firehose/CombiningFirehoseFactory.java
@@ -122,12 +122,6 @@ public class CombiningFirehoseFactory implements FirehoseFactory<InputRowParser>
     }
 
     @Override
-    public Runnable commit()
-    {
-      return currentFirehose.commit();
-    }
-
-    @Override
     public void close() throws IOException
     {
       currentFirehose.close();
