+++ server/src/main/java/org/apache/druid/segment/realtime/firehose/FixedCountFirehoseFactory.java
@@ -84,12 +84,6 @@ public class FixedCountFirehoseFactory implements FirehoseFactory
       }
 
       @Override
-      public Runnable commit()
-      {
-        return delegateFirehose.commit();
-      }
-
-      @Override
       public void close() throws IOException
       {
         delegateFirehose.close();
