+++ server/src/main/java/org/apache/druid/segment/realtime/firehose/IngestSegmentFirehose.java
@@ -44,7 +44,6 @@ import org.apache.druid.segment.data.IndexedInts;
 import org.apache.druid.segment.filter.Filters;
 import org.apache.druid.segment.transform.TransformSpec;
 import org.apache.druid.segment.transform.Transformer;
-import org.apache.druid.utils.Runnables;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
@@ -200,12 +199,6 @@ public class IngestSegmentFirehose implements Firehose
     final InputRow inputRow = rowYielder.get();
     rowYielder = rowYielder.next(null);
     return transformer.transform(inputRow);
-  }
-
-  @Override
-  public Runnable commit()
-  {
-    return Runnables.getNoopRunnable();
   }
 
   @Override
