+++ server/src/main/java/org/apache/druid/segment/realtime/firehose/PredicateFirehose.java
@@ -82,12 +82,6 @@ public class PredicateFirehose implements Firehose
   }
 
   @Override
-  public Runnable commit()
-  {
-    return firehose.commit();
-  }
-
-  @Override
   public void close() throws IOException
   {
     firehose.close();
